requirements.txt
# Requirements for AutoPatch-PR-Agent
# Core SDKs and HTTP
google-genai>=0.5.0
requests>=2.28

# Git operations
GitPython>=3.1.30

# Data validation
pydantic>=1.10

# Linting / formatting (Python)
ruff>=0.15.2
black>=23.9.0

# Local web UI
streamlit>=1.20

# Visualization
plotly>=5.0.0

# Optional helpers
python-dotenv>=1.0.0
bandit>=1.7.0
